BE Create PG & Connect/update DB, update: Routes
Created model for cars to connect to MySQL.
Moved some routes into the cars controller I created.
User controller now has only the things that pertain to users.
Updated the DB for cars to input all the data needed.
Fixed the spacing between the listings in the home page.
Updated the styling of the cards/containers in home/create pages.