Update: complete the FE of the edit_listing page and BE (validations & DB query update)
Updated the car validation static method function to skip the 'current_car_title' when looking for title duplicates in validations if it exists as the same input when you open up the page to edit a car listing.

Update error validation only appears now if the input for the title is not the same as the existing car listing title that's being edited & or a car listing title that already exists.

Fixed: update car class method is now working.