Update: the folder setup for static: individual css, js, images folders with respective files in them.
Update: the linking path to the css on all html templates. (all styling from css is working good)

Fixed: big where it wouldn't show the create_listing_page since it was looking for a 'current_car_title' session and it didn't exist since it got session.pop from the app.

Update: linking path to the javascript on all html templates. (the rendering for the image from the selected file input is working)